Migrate ssm
